requirements.txt
# Framework de Automatización QA para DemoBlaze - Dependencias Esenciales

# Selenium WebDriver
selenium==4.15.2
webdriver-manager==4.0.1

# Framework de testing
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-timeout==2.2.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-sugar==0.9.7

# Reportes
allure-pytest==2.13.2
allure-python-commons==2.13.2

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
